Purple Ball Draft
Able to make purple freeze, but velocity going down everytime it froze and no acceleration.

+++ src/main/java/org/example/BallPit.java
@@ -70,7 +70,7 @@ class BallPit {
                 }
             }
 
-            ball.think((int) tickCount);
+            ball.think(tickCount);
         }
     }
